Add a real time clock.
This commit adds the clock to the device tree, adds an rc.local that disables
the i2c driver's attempt to disable to 1Hz SQW output, adds udev rules to ensure
that the GPS and RTC have unique names in /dev, and has chrony monitor its
frequency.

The PPS stuff, chrony config, udev rules, etc. are not necessary to use the
device as the default hardware clock upon boot; that is all handled by the
device tree.