device tree: spi: reduce frequency
There is a rats nest of wires inside the clock, and the display glitched from time to time (on the
order of a few times a day).  Perhaps reducing the clock to 1MHz will help with this.  (20ms to
display the time now, which makes me a little sad.)

I also set the default SPI mode to Mode 3, which the APA102 datasheet says the LEDs expect.  The Go
code did set the bus to Mode 3, but it reverted to Mode 0 when idle.  This was never a problem, but
it seems like having the clock rise while the data line is low can look like a 0 bit, so keeping the
SPI bus in Mode 3 seems likely to reduce problems.  (Changing two variables at once... maybe this
was the real cause of glitches, but the glitches didn't look like a 1 bit offset, so probably not.)

I did observe the signal with the 'scope after makign this change and it does work.  Mode 3, 1MHz
now!